Add options to IO
To pass the original hook stdIn as --input option the
IO had to be made aware of the command options.
And all detectors da to switch from reading stdIn to asking
the IO for the --input option.

This way no hook execution is depending on stdIn and all hook
actions can require user input if deemed necessary.